requirements.txt
alembic==1.6.5
anyio==3.3.0
asgiref==3.4.1
async-exit-stack==1.0.1
async-generator==1.10
asyncpg==0.24.0
autopep8==1.5.7
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.1
cryptography==3.4.7
cssselect==1.1.0
cssutils==2.3.0
databases==0.4.3
Deprecated==1.2.12
dnspython==2.1.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.68.0
fastapi-users==6.1.2
flake8==3.9.2
greenlet==1.1.1
h11==0.12.0
httpcore==0.13.6
httpx==0.18.2
httpx-oauth==0.3.6
idna==3.2
lxml==4.6.3
makefun==1.11.3
Mako==1.1.4
MarkupSafe==2.0.1
mccabe==0.6.1
passlib==1.7.4
premailer==3.10.0
psycopg2==2.9.1
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pydantic==1.8.2
pyflakes==2.3.1
PyJWT==2.1.0
python-dateutil==2.8.2
python-decouple==3.4
python-editor==1.0.4
python-jose==3.3.0
python-multipart==0.0.5
requests==2.26.0
rfc3986==1.5.0
rsa==4.7.2
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.3.24
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
urllib3==1.26.6
uvicorn==0.14.0
wrapt==1.12.1
yagmail==0.14.256
